packages/joke/joke/requirements.txt
pyjokes==0.6.0
psycopg2-binary==2.9.6
